+++ sdks/python/clienthub/configuration.py
@@ -531,7 +531,7 @@ conf = clienthub.Configuration(
                "OS: {env}\n"\
                "Python Version: {pyversion}\n"\
                "Version of the API: 0.3.6\n"\
-               "SDK Package Version: 0.3.6".\
+               "SDK Package Version: 0.4.0".\
                format(env=sys.platform, pyversion=sys.version)
 
     def get_host_settings(self) -> List[HostSetting]: